refactor: 스켈레톤 loading파일이 아닌 일반파일로 수정 (#298)
