+++ app/(dashboard)/strategies/[strategyId]/_ui/details-skeleton/index.tsx
@@ -1,6 +1,6 @@
 import classNames from 'classnames/bind'
 
-import styles from './loading.module.scss'
+import styles from './styles.module.scss'
 
 const cx = classNames.bind(styles)
