+++ app/(dashboard)/strategies/[strategyId]/_ui/details-skeleton/styles.module.scss
