+++ app/(dashboard)/strategies/page.tsx
@@ -5,11 +5,11 @@ import classNames from 'classnames/bind'
 import Title from '@/shared/ui/title'
 
 import ListHeader from '../_ui/list-header'
+import StrategiesItemSkeleton from '../_ui/strategies-item/skeleton'
 import SearchBarContainer from './_ui/search-bar'
 import SearchBarSkeleton from './_ui/search-bar/search-bar-skeleton'
 import SideContainer from './_ui/side-container'
 import StrategyList from './_ui/strategy-list'
-import StrategiesLoading from './loading'
 import styles from './page.module.scss'
 
 const cx = classNames.bind(styles)
@@ -19,7 +19,7 @@ const StrategiesPage = () => {
     <div className={cx('container')}>
       <Title label={'전략 랭킹 모음'} />
       <ListHeader />
-      <Suspense fallback={<StrategiesLoading />}>
+      <Suspense fallback={<Skeleton />}>
         <StrategyList />
       </Suspense>
       <SideContainer>
@@ -28,6 +28,16 @@ const StrategiesPage = () => {
         </Suspense>
       </SideContainer>
     </div>
+  )
+}
+
+const Skeleton = () => {
+  return (
+    <>
+      {Array.from({ length: 6 }, (_, idx) => (
+        <StrategiesItemSkeleton key={idx} />
+      ))}
+    </>
   )
 }
